requirements.txt
customtkinter
opencv-python
tensorflow
psutil
Pillow
numpy
easyocr
playsound
